Database bağlantıları yapıldı
Database bağlantıları yapıldı .Senkronizasyon kısmında hata var (tekrar bakılmalı).ReviewPage ve FavoritePage üzerinden mekan sayfasına geçince navigasyon bar güncellenmiyor(düzeltilmeli).

+++ app/lib/services/favoritesservice.dart
@@ -26,7 +26,7 @@ class FavoritesService {
     required String userId,
     required String placeName,
     required String placeAddress,
-    required int rating,
+    required double rating,
     required String photoUrl,
   }) async {
     final uri = Uri.parse('$baseUrl/addFavorite');
